+++ CHANGELOG.md
@@ -4,6 +4,7 @@ walkers 8.0
 - disable swimming for some humanoid mobs
 - **port to 26.1.2** ~thanks to tgboyles
 - fix flying for Vex
+- fix the "There are Monsters Nearby" bed rule for disguised players
 
 walkers 7.2.7
 ================